Feature: send tts text to webhook
Since you may only have a microphone, you may want to send the text returned by the Wyoming Sattelite elsewhere to be spoken (ie. anohter media plaeyer). In this example the text is sent to a HA webhook so you can do lots of further automation. An example is provided in the DOCS.

This is the first, untested commit. It's likely to completely fail :-)